packages/wallet/src/functions/sidetreeEncoding.ts
/*
 * Copyright 2020 - Transmute Industries Inc.
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *     http://www.apache.org/licenses/LICENSE-2.0
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import crypto from 'crypto';
import base64url from 'base64url';
import multihashes from 'multihashes';
import canonicalize from 'canonicalize';

export const sha256 = (data: Buffer): Buffer => {
  return crypto.createHash('sha256').update(data).digest();
};

export const hashThenEncode = (data: Buffer): string => {
  const bytes = new Uint8Array(Buffer.from(sha256(data)));
  return base64url.encode(multihashes.encode(bytes, 'sha2-256'));
};

export const canonicalizeThenHashThenEncode = (data: object): string => {
  const cannonical = canonicalize(data);
  return hashThenEncode(cannonical);
};
